If you only look on the surface, all you will ever see is the snapshot.
Anyone who has worked on software knows that the state of the project in
question is a direct result of its history. When we want to understand why
something was done a certain way, we check the BLAME. When we want to
understand how the source has evolved over time, we check the commit
log. As people, we have our own internal git repository. We can replay
the memories and experiences that led to who we are today. But when we
look at others, we only see the snapshot, the binary, the artifact, the
working tree.

You are not the only one with an internal repository. You have to look
deeper.