I try my best to make people think of me as a good person.
I hate feigning interest, and in all honesty I feel like I spend most of
my day doing that. The polite nods, the courtesy chuckles, the
reciprocated smile...it feels like a lot of work.

But even more than I hate feigning interest, I hate the feeling of being
excluded or having a target on my back. So in general I've learned how
to be a fairly likable person even when I don't feel so great on the
inside.

But maybe I just need to learn to see the beauty in everything so that I
can genuinely find anything interesting. Is that realistic?